+++ at_commons/lib/src/shared_key_status.dart
@@ -1,9 +1,0 @@
-enum SharedKeyStatus {
-  LOCAL_UPDATED,
-  REMOTE_UPDATED,
-  SHARED_WITH_NOTIFIED,
-  SHARED_WITH_LOOKED_UP,
-  SHARED_WITH_READ
-}
-
-String getSharedKeyName(SharedKeyStatus d) => '$d'.split('.').last;